+++ CHANGELOG.md
@@ -8,6 +8,20 @@ and this project adheres to
 
 ## Unreleased
 
+### Added
+
+- New entity added:
+
+| Resources | Entity `_type` | Entity `_class` |
+| --------- | -------------- | --------------- |
+| Okta Rule | `okta_rule`    | `Configuration` |
+
+- New relationship added:
+
+| Source Entity `_type` | Relationship `_class` | Target Entity `_type` |
+| --------------------- | --------------------- | --------------------- |
+| `okta_account`        | **HAS**               | `okta_rule`           |
+
 ## 2.0.0 - 2021-06-22
 
 ### Added